fix!: git directory rename bug
