+++ resources/js/components/ApplicationLogo.tsx
