+++ resources/js/components/TextInput.tsx
